feat: Added Lsit view to present Hospital data in a historical context broken down by each day since the data was taken
The list view added has a list item  title of the current date for the information followed by total hospitalized with change in the past 24 hours added to that HStack, Total in ICU, and Total in Acute care for that given day. The data is taken from the Maryland State gov Open Data API for Covid 19 and called from the Observed object of CovidMasterModel class.