style: Changed the list view to present in a small card that can be tapped to enlarge
This styling was added in order to give the user a more animated view of the list with a small pop up that is centered in the middle of the screen. The background around the list card is blurred and the title aboce reflects the information beind shown in the list. The list can be pressed to enlarge the list view with a custom Spring animation