style: Added data to show on list card with animations of expansion on tap
This styling was added in order to have a list view that is in a card on tap of the list button in CaseCard. The list can be expanded when tapping and the background is blured for the entire Homeview upon activation to let the user focus on list data. There is an issue noted with a delay upon display of information from this api call which is hoped to be fixed by adding a loading Lottie animation on button tep