feat: Added two features of dismissing cardview on tap outside the card view, and added feature to dismiss card view on swipe with custom animation
In order to not activate any other buttons in the app while the card list is showing, a white sheet was added with a opacity modifer that takes up the background of the entire screen. On tap of this added sheet the card list is dismissed. The added swipe gesture was made with a custom animation that dimisses the view when touching outside the list in the list card and swiping right and down