feat: Added two features to this view. First, the dismissing of the list card when user taps outside of the list card when it is activated and two, a swipe gesture that dismisses the card with custom animation when the list card is showing by dragging outside of the list
The first feature was added in order to no activate any other button or swiping in the CaseView when the list card was showing. This was added by setting a opaque background to the lsit card that shows behind the list card in a Zstack and upon tapping, closes the list. The Swipe gesture added is on the BlurView of the list card that dismisses the view with a custom 3d rotation on swiping right and down